Refactor packaging wasm and gopherjs into reusing package web.
