+++ cmd/fyne/internal/commands/package-web.go
@@ -10,54 +10,57 @@ import (
 	"fyne.io/fyne/v2/cmd/fyne/internal/util"
 )
 
-func (p *Packager) packageWeb() error {
-	appDir := util.EnsureSubDir(p.dir, "web")
+type webData struct {
+	WasmFile     string
+	GopherJSFile string
+	IsReleased   bool
+	HasWasm      bool
+	HasGopherJS  bool
+}
 
+func packageWebInternal(w webData, appDir string, exeWasmSrc string, exeJSSrc string, icon string, release bool) error {
 	index := filepath.Join(appDir, "index.html")
 	indexFile, err := os.Create(index)
 	if err != nil {
 		return err
 	}
 
-	tplData := indexData{
-		GopherJSFile: p.name + ".js",
-		WasmFile:     p.name + ".wasm",
-		IsReleased:   p.release,
-		HasGopherJS:  true,
-		HasWasm:      true,
-	}
-	err = templates.IndexHTML.Execute(indexFile, tplData)
+	err = templates.IndexHTML.Execute(indexFile, w)
 	if err != nil {
 		return err
 	}
 
 	iconDst := filepath.Join(appDir, "icon.png")
-	err = util.CopyFile(p.icon, iconDst)
+	err = util.CopyFile(icon, iconDst)
 	if err != nil {
 		return err
 	}
 
-	exeJSDst := filepath.Join(appDir, p.name+".js")
-	err = util.CopyFile(p.exe+".js", exeJSDst)
-	if err != nil {
-		return err
+	if w.HasGopherJS {
+		exeJSDst := filepath.Join(appDir, w.GopherJSFile)
+		err = util.CopyFile(exeJSSrc, exeJSDst)
+		if err != nil {
+			return err
+		}
 	}
 
-	wasmExecSrc := filepath.Join(runtime.GOROOT(), "misc", "wasm", "wasm_exec.js")
-	wasmExecDst := filepath.Join(appDir, "wasm_exec.js")
-	err = util.CopyFile(wasmExecSrc, wasmExecDst)
-	if err != nil {
-		return err
-	}
+	if w.HasWasm {
+		wasmExecSrc := filepath.Join(runtime.GOROOT(), "misc", "wasm", "wasm_exec.js")
+		wasmExecDst := filepath.Join(appDir, "wasm_exec.js")
+		err = util.CopyFile(wasmExecSrc, wasmExecDst)
+		if err != nil {
+			return err
+		}
 
-	exeWasmDst := filepath.Join(appDir, p.name+".wasm")
-	err = util.CopyFile(p.exe+".wasm", exeWasmDst)
-	if err != nil {
-		return err
+		exeWasmDst := filepath.Join(appDir, w.WasmFile)
+		err = util.CopyFile(exeWasmSrc, exeWasmDst)
+		if err != nil {
+			return err
+		}
 	}
 
 	// Download webgl-debug.js directly from the KhronosGroup repository when needed
-	if !p.release {
+	if !release {
 		webglDebugFile := filepath.Join(appDir, "webgl-debug.js")
 		err := ioutil.WriteFile(webglDebugFile, templates.WebGLDebugJs, 0644)
 		if err != nil {
@@ -66,4 +69,42 @@ func (p *Packager) packageWeb() error {
 	}
 
 	return nil
+}
+
+func (p *Packager) packageWeb() error {
+	appDir := util.EnsureSubDir(p.dir, "web")
+
+	tpl := webData{
+		GopherJSFile: p.name + ".js",
+		WasmFile:     p.name + ".wasm",
+		IsReleased:   p.release,
+		HasGopherJS:  true,
+		HasWasm:      true,
+	}
+
+	return packageWebInternal(tpl, appDir, p.exe+".wasm", p.exe+".js", p.icon, p.release)
+}
+
+func (p *Packager) packageWasm() error {
+	appDir := util.EnsureSubDir(p.dir, "wasm")
+
+	tpl := webData{
+		WasmFile:   p.name,
+		IsReleased: p.release,
+		HasWasm:    true,
+	}
+
+	return packageWebInternal(tpl, appDir, p.exe, "", p.icon, p.release)
+}
+
+func (p *Packager) packageGopherJS() error {
+	appDir := util.EnsureSubDir(p.dir, "gopherjs")
+
+	tpl := webData{
+		GopherJSFile: p.name,
+		IsReleased:   p.release,
+		HasGopherJS:  true,
+	}
+
+	return packageWebInternal(tpl, appDir, "", p.exe, p.icon, p.release)
 }
